Pre-release master branch Github repo.
The Excel file is provided with the .bas files already imported and ready-to-use. They are listed here only for convenience to anyone who wishes to view the code. As-is, the Excel is incomplete, entirely missing the Flares section and also missing any means of automating ammo-kill assignments: to track ammo, users must manually calculate relevant formulas/values. Also, many ranges that ought to be are not set to protected, meaning the user could very easily "break" something with innocuous clicks or typing.